+++ biz/dal/db/mysql/ent/schema/logs.go
@@ -24,7 +24,7 @@ func (Logs) Fields() []ent.Field {
 		field.String("ip").Optional().Comment("ip of log | 日志IP"),
 		field.String("user_agent").Optional().Comment("user_agent of log | 日志用户客户端"),
 		field.String("operator").Optional().Comment("operator of log | 日志操作者"),
-		field.Int("time").Optional().Comment("time of log(millisecond) | 日志时间(毫秒)"),
+		field.Int64("time").Optional().Comment("time of log(millisecond) | 日志时间(毫秒)"),
 	}
 }
